+++ SymbolBuilder/Readers/SchLibPinReader.cs
@@ -14,12 +14,22 @@ namespace SymbolBuilder.Readers
         public override bool CanRead(string fileName) =>
             Path.GetExtension(fileName).Equals(".schlib", StringComparison.InvariantCultureIgnoreCase);
 
-        public override List<Package> LoadFromStream(Stream stream)
+        public override List<Package> LoadFromStream(Stream stream, string fn = null)
         {
             var fileName = Path.GetTempFileName();
-            using (var fs = File.OpenWrite(fileName))
+            bool createdTempFile = false;
+
+            if (string.IsNullOrEmpty(fn))
             {
-                stream.CopyTo(fs);
+                using (var fs = File.OpenWrite(fileName))
+                {
+                    stream.CopyTo(fs);
+                    createdTempFile = true;
+                }
+            }
+            else
+            {
+                fileName = fn;
             }
 
             var reader = new AltiumSharp.SchLibReader();
@@ -36,7 +46,10 @@ namespace SymbolBuilder.Readers
                 result.Add(package);
             }
 
-            File.Delete(fileName);
+            if (createdTempFile)
+            {
+                File.Delete(fileName);
+            }
 
             return result;
         }
